move validator files to correct package (data) since it handles business logic
